+++ NEWS.md
@@ -8,6 +8,7 @@ This is an update that corresponds to Python Polars 1.32.2.
 
 - New function `arr$len()` (#1478).
 - New function `<expr>$str$normalize()` (#1479).
+- New function `<expr>$str$find_many()` (#1483).
 
 ## polars 1.1.0